# ajpt/swagger/swagger_output.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "JackPote API",
      "description": "Documentation for JackPote API."
    },
    "host": "51.77.141.175:3000",
    "basePath": "/",
    "schemes": [
      "http",
      "https"
    ],
    "securityDefinitions": {
      "Bearer": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Rentrez un token valide dans le format **Bearer &lt;token>**"
      }
    },
    "produces": [
      "application/json"
    ],
    "paths": {
...
      "/object_profile/create/init": {
        "post": {
          "tags": [
            "Object Profile"
          ],
          "description": "Création d'object_profiles.",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "File",
              "in": "body",
              "description": "Champs object_profiles",
              "schema": {
                "$ref": "#/definitions/ObjectProfileCreate"
              }
            }
          ],
          "responses": {},
          "security": [
            {
              "Bearer": []
            }
          ]
        }
      },
      "/object_profile/resume/byperson": {
        "get": {
          "tags": [
            "Object Profile"
          ],
          "summary": "Résumé des profils d'objets par personne",
          "description": "Récupère la liste des objets activés appartenant à l'utilisateur connecté via son token (Bearer).",
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "Liste des profils d'objets récupérée avec succès",
              "schema": {
                "type": "object",
                "properties": {
                  "status": { "type": "string", "example": "OK" },
                  "message": { "type": "string", "nullable": true, "example": null },
                  "code": { "type": "integer", "example": 200 },
                  "data": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id_object_profile": { "type": "integer", "example": 2 },
                        "title": { "type": "string", "example": "Ma Plante Salon" },
                        "is_automatic": { "type": "boolean", "example": false },
                        "is_water": { "type": "boolean", "example": true },
                        "state": { "type": "integer", "example": 5 },
                        "plant_type": {
                          "type": "object",
                          "properties": {
                            "id_plant_type": { "type": "integer", "example": 1 },
                            "title": { "type": "string", "example": "Lavande" },
                            "path_picture": { "type": "string", "example": "dataset/plants/lavande_ref.png" },
                            "avatar": {
                              "type": "object",
                              "nullable": true,
                              "properties": {
                                "title": { "type": "string", "example": "Lavande Heureuse" },
                                "path_picture": { "type": "string", "example": "dataset/avatar/Lavande/Lavande5.png" }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Non autorisé - Token absent ou invalide"
            },
            "500": {
              "description": "Erreur interne du serveur"
            }
          },
          "security": [
            {
              "Bearer": []
            }
          ]
        }
      },
      "/object_profile/resume/favoris/byperson": {
        "get": {
          "tags": [
            "Object Profile"
          ],
          "summary": "Résumé des profils favoris",
          "description": "Récupère la liste des objets marqués comme favoris appartenant à l'utilisateur connecté via son token.",
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "Liste des favoris récupérée avec succès",
              "schema": {
                "type": "object",
                "properties": {
                  "status": { "type": "string", "example": "OK" },
                  "code": { "type": "integer", "example": 200 },
                  "data": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id_object_profile": { "type": "integer", "example": 3 },
                        "title": { "type": "string", "example": "Ma Plante Préférée" },
                        "is_automatic": { "type": "boolean", "example": false },
                        "is_water": { "type": "boolean", "example": false },
                        "state": { "type": "integer", "example": 1 },
                        "path_picture": { "type": "string", "example": "uploads/ma_plante.jpg" },
                        "plant_type": {
                          "type": "object",
                          "properties": {
                            "id_plant_type": { "type": "integer", "example": 1 },
                            "title": { "type": "string", "example": "Lavande" },
                            "avatar": {
                              "type": "object",
                              "nullable": true,
                              "properties": {
                                "title": { "type": "string", "example": "Lavande Heureuse" },
                                "path_picture": { "type": "string", "example": "dataset/avatar/Lavande/Lavande0.png" }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Non autorisé - Token manquant ou invalide"
            },
            "500": {
              "description": "Erreur interne du serveur"
            }
          },
          "security": [
            {
              "Bearer": []
            }
          ]
        }
      },
      "/object_profile/update/byobjectprofile": {
        "patch": {
          "tags": [
            "Object Profile"
          ],
          "description": "modifier les object_profiles.",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "op",
              "in": "body",
              "description": "Listes object_profiles summary",
              "schema": {
                "$ref": "#/definitions/ObjectProfileUP"
              }
            }
          ],
          "responses": {},
          "security": [
            {
              "Bearer": []
            }
          ]
        }
      },
      "/object_profile/detail/byop": {
        "post": {
          "tags": [
            "Object Profile"
          ],
          "description": "lister le résumé des favoris des object_profiles.",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "op",
              "in": "body",
              "description": "Listes object_profiles summary",
              "schema": {
                "$ref": "#/definitions/ObjectProfileDetails"
              }
            }
          ],
          "responses": {},
          "security": [
            {
              "Bearer": []
            }
          ]
        }
      },
      "/object_profile/delete/byobjectprofile": {
        "delete": {
          "tags": [
            "Object Profile"
          ],
          "description": "Supprimer un object_profiles.",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "op",
              "in": "body",
              "description": "Del object_profiles",
              "schema": {
                "$ref": "#/definitions/ObjectProfileDel"
              }
            }
          ],
          "responses": {},
          "security": [
            {
              "Bearer": []
            }
          ]
        }
      },
...
    },
    "definitions": {
...
      "ObjectProfileCreate": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "NamePlant"
          },
          "id_object": {
            "type": "int",
            "example": 1
          },
          "id_plant_type": {
            "type": "int",
            "example": 1
          },
          "id_person": {
            "type": "int",
            "example": 19
          },
          "activate": {
            "type": "int",
            "example": 0
          }
        }
      },
...
      "ObjectProfileDel": {
        "type": "object",
        "properties": {
            "id_person": {
              "type": "int",
              "example": "1"
            },
            "id_object_profile": {
              "type": "int",
              "example": "1"
            }
          }
        },
        "ObjectProfileDetails": {
          "type": "object",
          "properties": {
              "id_object_profile": {
                "type": "int",
                "example": "1"
              }
            }
        },
        "ObjectProfileUP": {
          "type": "object",
          "properties": {
            "id_person": {
              "type": "int",
              "example": "1"
            },
            "id_object_profile": {
              "type": "int",
              "example": "1"
            }
            }
        },
...
    }
  }